+++ src/components/NavBar.jsx
@@ -1,21 +1,35 @@
-import React from 'react';
+import React, { useState }  from 'react';
 import { Link } from 'react-router-dom';
-import './NavBar.css'; 
+import './NavBar.css';
 
 function NavBar() {
-  return (
-    <nav>
-        <h1>Captains Log</h1>
-      <ul>
-        <li>
-          <Link to="/logs">All Logs</Link>
-        </li>
-        <li>
-          <Link to="/logs/new">New Log</Link>
-        </li>
-      </ul>
-    </nav>
-  );
+    // State to track dark mode
+    const [isDarkMode, setIsDarkMode] = useState(false);
+
+    // Function to toggle dark mode
+    const toggleDarkMode = () => {
+        const body = document.body;
+        body.classList.toggle('dark-mode');
+        setIsDarkMode(!isDarkMode); // Update state to track dark mode
+    };
+
+    return (
+
+        <nav>
+            <button id="dark-mode-toggle" onClick={toggleDarkMode}>
+                Toggle Dark Mode
+            </button>
+            <h1>Captains Log</h1>
+            <ul>
+                <li>
+                    <Link to="/logs">All Logs</Link>
+                </li>
+                <li>
+                    <Link to="/logs/new">New Log</Link>
+                </li>
+            </ul>
+        </nav>
+    );
 }
 
 export default NavBar;
